code refactored with the recomendations to the date
