+++ lib/credential_lib/credentialLib.js
@@ -1,5 +1,5 @@
-var login = require('../lib/login_libs/loginLib.js');
-var config = require('../config.json');
+var login = require('../login_libs/loginLib.js');
+var config = require('../../config.json');
 
 module.exports.getCredentialsJson = function () {
   var credentialsJson = {
@@ -14,4 +14,4 @@ module.exports.token = function (callback) {
     var token = config.token + res.body.token;
     callback(token);
   });
-}
+};
