+++ lib/login_libs/loginLib.js
@@ -1,5 +1,5 @@
-var request = require('../request_client/request.js');
-var resource = require('../../utilities/resource.json');
+var request = require('../request_lib/requestLib.js');
+var resource = require('../../resources/resource.json');
 
 module.exports.create = function (credentialsJson, callback) {
 	var endPoint = resource.login;
