+++ lib/request_lib/requestAutenticationLib.js
@@ -1,8 +1,8 @@
 var request = require('superagent');
 var config = require('../../config.json');
-var utility = require('../../utilities/utility.json');
+var nodeConfig = require('../../resources/node.config.json');
 
-process.env.NODE_TLS_REJECT_UNAUTHORIZED = utility.NODE_TLS_REJECT_UNAUTHORIZED;
+process.env.NODE_TLS_REJECT_UNAUTHORIZED = nodeConfig.NODE_TLS_REJECT_UNAUTHORIZED;
 
 module.exports.get = function (endPoint, callback) {
   request
