+++ lib/request_lib/requestTokenLib.js
@@ -1,9 +1,9 @@
 var request = require('superagent');
 var config = require('../../config.json');
-var utility = require('../../utilities/utility.json');
-var credentials = require('../../utilities/credentials.js');
+var credentials = require('../credential_lib/credentialLib.js');
+var nodeConfig = require('../../resources/node.config.json');
 
-process.env.NODE_TLS_REJECT_UNAUTHORIZED = utility.NODE_TLS_REJECT_UNAUTHORIZED;
+process.env.NODE_TLS_REJECT_UNAUTHORIZED = nodeConfig.NODE_TLS_REJECT_UNAUTHORIZED;
 
 module.exports.get = function (endPoint, callback) {
   credentials.token(function (token) {
