+++ resources/node.config.json
